+++ lib/src/models/phone_number_impl.dart
@@ -1,6 +1,9 @@
+import 'package:phone_number_metadata/phone_number_metadata.dart';
 import 'package:phone_number_metadata/src/models/phone_metadata.dart';
+import 'package:phone_numbers_parser/src/parsers/_validator.dart';
 
 import 'phone_number.dart';
+import 'phone_number_type.dart';
 
 /// factory pattern with the parsers being the factories
 class PhoneNumberImpl implements PhoneNumber {
@@ -20,6 +23,22 @@ class PhoneNumberImpl implements PhoneNumber {
   String get international => '+' + dialCode + nsn;
 
   const PhoneNumberImpl(this.nsn, this.metadata);
+
+  /// Validates a phone number using pattern matching
+  ///
+  /// if a [type] is added, will validate against a specific type
+  @override
+  bool validate([PhoneNumberType? type]) {
+    if (metadata is PhoneMetadataExtended) {
+      return Validator.validateWithPattern(
+        nsn,
+        metadata as PhoneMetadataExtended,
+        type,
+      );
+    } else {
+      return Validator.validateWithLength(nsn, metadata, type);
+    }
+  }
 
   @override
   bool operator ==(Object other) {