+++ test/phone_parser_test.dart
@@ -75,24 +75,5 @@ void main() {
       expect(
           parse('+54 9 343 555 1212').international, equals('+5493435551212'));
     });
-
-    test('should validate with pattern', () {
-      final validate = parser.validate;
-      final validMobile = parser.parseWithIsoCode('BE', '479 99 99 99');
-      final validCA = parser.parseWithIsoCode('CA', '+16135550165');
-      final validUS = parser.parseWithIsoCode('US', '+12025550128');
-      final invalidCA = parser.parseWithIsoCode('US', '+16135550165');
-      final invalidUS = parser.parseWithIsoCode('CA', '+12025550128');
-
-      // belgian phone numbers dont have the same length for fixedLine vs mobile
-      expect(validate(validMobile), equals(true));
-      expect(validate(validMobile, PhoneNumberType.mobile), equals(true));
-      expect(validate(validMobile, PhoneNumberType.fixedLine), equals(false));
-      // canadian and us can have same length numbers but the patterns differ
-      expect(validate(validUS), equals(true));
-      expect(validate(validCA), equals(true));
-      expect(validate(invalidUS), equals(false));
-      expect(validate(invalidCA), equals(false));
-    });
   });
 }